Move HTTP adapter base classes to HTTP adapter module
There are several classes in the service-base module which are only
relevant to HTTP based protocol adapters. These classes have therefore
been moved to the HTTP adapter base module.
