+++ adapters/http-vertx-base/src/main/java/org/eclipse/hono/adapter/http/HonoChainAuthHandler.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2019, 2020 Contributors to the Eclipse Foundation
+ * Copyright (c) 2019, 2021 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -12,11 +12,12 @@
  */
 
 
-package org.eclipse.hono.service.http;
+package org.eclipse.hono.adapter.http;
 
 import org.eclipse.hono.client.ServiceInvocationException;
 import org.eclipse.hono.service.auth.device.DeviceCredentialsAuthProvider;
 import org.eclipse.hono.service.auth.device.PreCredentialsValidationHandler;
+import org.eclipse.hono.service.http.HttpContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
