+++ services/device-registry-jdbc/src/main/java/org/eclipse/hono/deviceregistry/jdbc/ApplicationConfig.java
@@ -55,7 +55,6 @@ import org.eclipse.hono.service.base.jdbc.store.tenant.ManagementStore;
 import org.eclipse.hono.service.base.jdbc.store.tenant.Stores;
 import org.eclipse.hono.service.credentials.CredentialsService;
 import org.eclipse.hono.service.credentials.DelegatingCredentialsAmqpEndpoint;
-import org.eclipse.hono.service.http.HonoBasicAuthHandler;
 import org.eclipse.hono.service.http.HttpEndpoint;
 import org.eclipse.hono.service.http.HttpServiceConfigProperties;
 import org.eclipse.hono.service.management.credentials.CredentialsManagementService;
@@ -90,6 +89,7 @@ import io.vertx.ext.auth.AuthProvider;
 import io.vertx.ext.auth.jdbc.JDBCAuth;
 import io.vertx.ext.jdbc.JDBCClient;
 import io.vertx.ext.web.handler.AuthHandler;
+import io.vertx.ext.web.handler.BasicAuthHandler;
 
 /**
  * Spring Boot configuration for the JDBC based device registry application.
@@ -590,7 +590,7 @@ public class ApplicationConfig {
      * Creates a new instance of an auth handler to provide basic authentication for the 
      * HTTP based Device Registry Management endpoint.
      * <p>
-     * This creates an instance of the {@link HonoBasicAuthHandler} with the auth provider returned by
+     * This method creates a {@link BasicAuthHandler} using the auth provider returned by
      * {@link #authProvider()} if the property corresponding to {@link HttpServiceConfigProperties#isAuthenticationRequired()}
      * is set to {@code true}.
      *
@@ -604,10 +604,9 @@ public class ApplicationConfig {
     @Scope("prototype")
     public AuthHandler createAuthHandler(final HttpServiceConfigProperties httpServiceConfigProperties) {
         if (httpServiceConfigProperties != null && httpServiceConfigProperties.isAuthenticationRequired()) {
-            return new HonoBasicAuthHandler(
+            return BasicAuthHandler.create(
                     authProvider(),
-                    httpServerProperties().getRealm(),
-                    tracer());
+                    httpServerProperties().getRealm());
         }
         return null;
     }
